api_crowdfunding/requirements.txt
Django==1.11.2
djangorestframework==3.6.3
djangorestframework_bulk
